removed CreateWallet utility method
