+++ protocol/xmrtaker/instance.go
@@ -101,18 +101,23 @@ func (inst *Instance) createOngoingSwap(s *swap.Info) error {
 		}
 
 		// TODO: do we want to transfer this back to the original account?
-		addr, err := monero.CreateWallet(
-			"xmrmaker-swap-wallet",
-			inst.backend.Env(),
-			inst.backend.XMRClient(),
-			kp,
+		conf := inst.backend.XMRClient().CreateABWalletConf()
+		abWalletCli, err := monero.CreateSpendWalletFromKeys(
+			conf,
+			kp, // TODO: Fix the key here?
 			s.MoneroStartHeight,
 		)
 		if err != nil {
 			return err
 		}
+		if inst.transferBack {
+			defer abWalletCli.CloseAndRemoveWallet()
+			// TODO: Do transfer here after key issue are fixed
+		} else {
+			defer abWalletCli.Close() // leave the wallet in place, as funds were not transferred back
+		}
 
-		log.Infof("refunded XMR from swap %s: wallet addr is %s", s.ID, addr)
+		log.Infof("refunded XMR from swap %s: wallet addr is %s", s.ID, abWalletCli.PrimaryAddress())
 		return nil
 	}
